+++ library/include/rocrand/rocrand.hpp
@@ -1591,8 +1591,13 @@ constexpr typename mtgp32_engine<DefaultSeed>::seed_type mtgp32_engine<DefaultSe
 
 /// \brief Random number engine based on the LFSR113 algorithm.
 ///
-/// lfsr113_engine is a random number engine based on the LFSR113 algorithm
-template<unsigned long long DefaultSeed = ROCRAND_LFSR113_DEFAULT_SEED>
+/// lfsr113_engine is an implementation of LFSR113 pseudorandom number generator,
+/// which is a linear feedback shift resgisters (LFSR) based generator created by Pierre L'Ecuyer.
+/// It produces random 32-bit \p unsigned \p int values on the interval [0; 2^32 - 1].
+template<unsigned int DefaultSeedX = ROCRAND_LFSR113_DEFAULT_SEED_X,
+         unsigned int DefaultSeedY = ROCRAND_LFSR113_DEFAULT_SEED_Y,
+         unsigned int DefaultSeedZ = ROCRAND_LFSR113_DEFAULT_SEED_Z,
+         unsigned int DefaultSeedW = ROCRAND_LFSR113_DEFAULT_SEED_W>
 class lfsr113_engine
 {
 public:
@@ -1601,9 +1606,10 @@ public:
     /// \copydoc philox4x32_10_engine::offset_type
     typedef unsigned long long offset_type;
     /// \copydoc philox4x32_10_engine::seed_type
-    typedef unsigned long long seed_type;
+    typedef uint4 seed_type;
     /// \copydoc philox4x32_10_engine::default_seed
-    static constexpr seed_type default_seed = DefaultSeed;
+    static constexpr seed_type default_seed
+        = {DefaultSeedX, DefaultSeedY, DefaultSeedZ, DefaultSeedW};
 
     /// \brief Constructs the pseudo-random number engine.
     ///
@@ -1612,18 +1618,14 @@ public:
     /// \param seed_value - seed value to use in the initialization of the internal state, see also seed()
     ///
     /// See also: hiprandCreateGenerator()
-    lfsr113_engine(seed_type seed_value = DefaultSeed,
-                   offset_type offset_value = 0)
+    lfsr113_engine(seed_type seed_value = {DefaultSeedX, DefaultSeedY, DefaultSeedZ, DefaultSeedW})
     {
         rocrand_status status;
         status = rocrand_create_generator(&m_generator, this->type());
-        if(status != ROCRAND_STATUS_SUCCESS) throw rocrand_cpp::error(status);
+        if(status != ROCRAND_STATUS_SUCCESS)
+            throw rocrand_cpp::error(status);
         try
         {
-            if (offset_value > 0)
-            {
-                this->offset(offset_value);
-            }
             this->seed(seed_value);
         }
         catch(...)
@@ -1634,8 +1636,7 @@ public:
     }
 
     /// \copydoc philox4x32_10_engine::philox4x32_10_engine(rocrand_generator&)
-    lfsr113_engine(rocrand_generator& generator)
-        : m_generator(generator)
+    lfsr113_engine(rocrand_generator& generator) : m_generator(generator)
     {
         if(generator == NULL)
         {
@@ -1648,30 +1649,42 @@ public:
     ~lfsr113_engine() noexcept(false)
     {
         rocrand_status status = rocrand_destroy_generator(m_generator);
-        if(status != ROCRAND_STATUS_SUCCESS) throw rocrand_cpp::error(status);
+        if(status != ROCRAND_STATUS_SUCCESS)
+            throw rocrand_cpp::error(status);
     }
 
     /// \copydoc philox4x32_10_engine::stream()
     void stream(hipStream_t value)
     {
         rocrand_status status = rocrand_set_stream(m_generator, value);
-        if(status != ROCRAND_STATUS_SUCCESS) throw rocrand_cpp::error(status);
+        if(status != ROCRAND_STATUS_SUCCESS)
+            throw rocrand_cpp::error(status);
+    }
+
+    /// \copydoc philox4x32_10_engine::seed()
+    void seed(unsigned long long value)
+    {
+        rocrand_status status = rocrand_set_seed(this->m_generator, value);
+        if(status != ROCRAND_STATUS_SUCCESS)
+            throw rocrand_cpp::error(status);
     }
 
     /// \copydoc philox4x32_10_engine::seed()
     void seed(seed_type value)
     {
-        rocrand_status status = rocrand_set_seed(this->m_generator, value);
-        if(status != ROCRAND_STATUS_SUCCESS) throw rocrand_cpp::error(status);
+        rocrand_status status = rocrand_set_seed_uint4(this->m_generator, value);
+        if(status != ROCRAND_STATUS_SUCCESS)
+            throw rocrand_cpp::error(status);
     }
 
     /// \copydoc philox4x32_10_engine::operator()()
     template<class Generator>
-    void operator()(result_type * output, size_t size)
+    void operator()(result_type* output, size_t size)
     {
         rocrand_status status;
         status = rocrand_generate(m_generator, output, size);
-        if(status != ROCRAND_STATUS_SUCCESS) throw rocrand_cpp::error(status);
+        if(status != ROCRAND_STATUS_SUCCESS)
+            throw rocrand_cpp::error(status);
     }
 
     /// \copydoc philox4x32_10_engine::min()
@@ -1689,7 +1702,7 @@ public:
     /// \copydoc philox4x32_10_engine::type()
     static constexpr rocrand_rng_type type()
     {
-        return ROCRAND_RNG_PSEUDO_MTGP32;
+        return ROCRAND_RNG_PSEUDO_LFSR113;
     }
 
 private:
@@ -1714,8 +1727,12 @@ private:
 };
 
 /// \cond
-template<unsigned long long DefaultSeed>
-constexpr typename lfsr113_engine<DefaultSeed>::seed_type lfsr113_engine<DefaultSeed>::default_seed;
+template<unsigned int DefaultSeedX,
+         unsigned int DefaultSeedY,
+         unsigned int DefaultSeedZ,
+         unsigned int DefaultSeedW>
+constexpr typename lfsr113_engine<DefaultSeedX, DefaultSeedY, DefaultSeedZ, DefaultSeedW>::seed_type
+    lfsr113_engine<DefaultSeedX, DefaultSeedY, DefaultSeedZ, DefaultSeedW>::default_seed;
 /// \endcond
 
 /// \brief Sobol's quasi-random sequence generator
@@ -2058,7 +2075,7 @@ typedef mrg32k3a_engine<> mrg32k3a;
 /// \brief Typedef of rocrand_cpp::mtgp32_engine PRNG engine with default seed (0).
 typedef mtgp32_engine<> mtgp32;
 /// \typedef lfsr113
-/// \brief Typedef of rocrand_cpp::lfsr113_engine PRNG engine with default seed (#ROCRAND_LFSR113_DEFAULT_SEE).
+/// \brief Typedef of rocrand_cpp::lfsr113_engine PRNG engine with default seed (#ROCRAND_LFSR113_DEFAULT_SEED).
 typedef lfsr113_engine<> lfsr113;
 /// \typedef sobol32
 /// \brief Typedef of rocrand_cpp::sobol32_engine PRNG engine with default number of dimensions (1).