+++ library/include/rocrand/rocrand_normal.h
@@ -917,9 +917,32 @@ double rocrand_normal_double(rocrand_state_sobol64 * state)
  * \return Normally distributed \p double value
  */
 FQUALIFIERS
-float rocrand_normal(rocrand_state_lfsr113 * state)
+float rocrand_normal(rocrand_state_lfsr113* state)
 {
     return rocrand_device::detail::normal_distribution_double(rocrand(state));
+}
+
+/**
+ * \brief Returns two normally distributed \p float values.
+ *
+ * Generates and returns two normally distributed \p float values using LFSR113
+ * generator in \p state, and increments position of the generator by two.
+ * Used normal distribution has mean value equal to 0.0f, and standard deviation
+ * equal to 1.0f.
+ * The function uses the Box-Muller transform method to generate two normally
+ * distributed values, and returns both of them.
+ *
+ * \param state - Pointer to a state to use
+ *
+ * \return Two normally distributed \p float value as \p float2
+ */
+FQUALIFIERS
+float2 rocrand_normal2(rocrand_state_lfsr113* state)
+{
+    auto state1 = rocrand(state);
+    auto state2 = rocrand(state);
+
+    return rocrand_device::detail::normal_distribution2(state1, state2);
 }
 
 /**
@@ -935,9 +958,35 @@ float rocrand_normal(rocrand_state_lfsr113 * state)
  * \return Normally distributed \p double value
  */
 FQUALIFIERS
-double rocrand_normal_double(rocrand_state_lfsr113 * state)
+double rocrand_normal_double(rocrand_state_lfsr113* state)
 {
     return rocrand_device::detail::normal_distribution_double(rocrand(state));
+}
+
+/**
+ * \brief Returns two normally distributed \p double values.
+ *
+ * Generates and returns two normally distributed \p double values using LFSR113
+ * generator in \p state, and increments position of the generator by two.
+ * Used normal distribution has mean value equal to 0.0f, and standard deviation
+ * equal to 1.0f.
+ * The function uses the Box-Muller transform method to generate two normally
+ * distributed values, and returns both of them.
+ *
+ * \param state - Pointer to a state to use
+ *
+ * \return Two normally distributed \p double value as \p double2
+ */
+FQUALIFIERS
+double2 rocrand_normal_double2(rocrand_state_lfsr113* state)
+{
+    auto state1 = rocrand(state);
+    auto state2 = rocrand(state);
+    auto state3 = rocrand(state);
+    auto state4 = rocrand(state);
+
+    return rocrand_device::detail::normal_distribution_double2(
+        uint4{state1, state2, state3, state4});
 }
 
 /** @} */ // end of group rocranddevice