+++ library/src/rng/generator_type.hpp
@@ -22,6 +22,7 @@
 #define ROCRAND_RNG_GENERATOR_TYPE_H_
 
 #include <hip/hip_runtime.h>
+#include <hip/hip_vector_types.h>
 #include <rocrand/rocrand.h>
 
 struct rocrand_generator_base_type
@@ -33,16 +34,16 @@ struct rocrand_generator_base_type
 };
 
 // rocRAND random number generator base class
-template<rocrand_rng_type GeneratorType = ROCRAND_RNG_PSEUDO_PHILOX4_32_10>
+template<rocrand_rng_type GeneratorType = ROCRAND_RNG_PSEUDO_PHILOX4_32_10,
+         class SeedType                 = unsigned long long>
 struct rocrand_generator_type : public rocrand_generator_base_type
 {
     using base_type = rocrand_generator_base_type;
 
-    rocrand_generator_type(unsigned long long seed = 0,
+    rocrand_generator_type(SeedType           seed   = SeedType{0},
                            unsigned long long offset = 0,
-                           hipStream_t stream = 0)
-        : base_type(GeneratorType),
-          m_seed(seed), m_offset(offset), m_stream(stream)
+                           hipStream_t        stream = 0)
+        : base_type(GeneratorType), m_seed(seed), m_offset(offset), m_stream(stream)
     {
 
     }
@@ -53,7 +54,7 @@ struct rocrand_generator_type : public rocrand_generator_base_type
         return rng_type;
     }
 
-    unsigned long long get_seed() const
+    SeedType get_seed() const
     {
         return m_seed;
     }
@@ -75,7 +76,7 @@ struct rocrand_generator_type : public rocrand_generator_base_type
 
 protected:
     // ordering type
-    unsigned long long m_seed;
+    SeedType           m_seed;
     unsigned long long m_offset;
     hipStream_t m_stream;
 };